requirements.txt
boto3==1.14.8
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Caching==1.9.0
Flask-Sessionstore==0.4.5
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
passlib[argon2]==1.7.2
psycopg2-binary==2.8.5
requests==2.24.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
ujson==3.0.0
python-dotenv==0.14.0
itsdangerous==1.1.0
bidict==0.21.2

# for checks
pytest-cov==2.10.0
black==19.10b0
flake8==3.8.3
